requirements.txt
langchain
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
streamlit
python-dotenv